3D login screen
this project is like this cool login page, right? It's not your average one, though. It's got this super awesome 3D space background with all these neat animations that really jazz it up. We put it together using HTML and CSS to make everything look nice and tidy, and we used JavaScript to make it all interactive and check the login details. Oh, and we threw in Three.js to make the background 3D and give everything a fun vibe. It's pretty rad, if you ask me!